PS-87 Transactions

+++ src/server/Persistify.Server.Management.Abstractions/Types/ITypeManagerHit.cs
@@ -2,5 +2,5 @@
 
 public interface ITypeManagerHit
 {
-    long DocumentId { get; }
+    int DocumentId { get; }
 }
