+++ src/server/Persistify.Server.Management/Number/Queries/EqualNumberManagerQuery.cs
@@ -17,7 +17,7 @@ public class EqualNumberManagerQuery : NumberManagerQuery
         Value = value;
     }
 
-    public override async ValueTask<List<NumberManagerHit>> Evaluate(IAsyncLookup<double, long> lookup)
+    public override async ValueTask<List<NumberManagerHit>> Evaluate(IAsyncLookup<double, int> lookup)
     {
         var documentIds = await lookup.GetAsync(Value);
         var hits = new List<NumberManagerHit>(documentIds.Count);
