+++ src/server/Persistify.Server.Management/Number/Queries/GreaterThanOrEqualNumberManagerQuery.cs
@@ -19,7 +19,7 @@ public class GreaterThanOrEqualNumberManagerQuery : NumberManagerQuery
         Value = value;
     }
 
-    public override async ValueTask<List<NumberManagerHit>> Evaluate(IAsyncLookup<double, long> lookup)
+    public override async ValueTask<List<NumberManagerHit>> Evaluate(IAsyncLookup<double, int> lookup)
     {
         var documentIds = await lookup.GetRangeAsync(Value, double.MaxValue);
         var hits = new List<NumberManagerHit>(documentIds.Count);
